Move setSchema to utils.js
Doesn't need to be in the prototype.
